02._Data_Files/pyproject.toml
[tool.poetry]
name = "02-data-files"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "02"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
requests = "^2.31.0"
uvicorn = "^0.27.1"
pyyaml = "^6.0.1"
xmltodict = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
